requirements.txt
fastapi[all]
uvicorn
motor[asyncio]
pyjwt
pydantic
passlib
pytest
pytest-asyncio
python-decouple
